Remove guard for evil-want-C-i-jump
It could be that someone bound evil-jump-forward to another key and then
wants to use evil-jumper in that case evil-want-C-i-jump would be nil
but evil-jump-forward is bound somewhere else. By removing the guard
this binding will automatically get remapped without extra effort.

When evil-jump-forward is not bound anywhere, the remap will not take
effect anyway, so it's harmless to remove the guard in this case.